+++ Ticketing.Infrastructure.EFCore/Migrations/TicketingContextModelSnapshot.cs
@@ -84,8 +84,8 @@ namespace Ticketing.Infrastructure.EFCore.Migrations
                         .HasMaxLength(50)
                         .HasColumnType("nvarchar(50)");
 
-                    b.Property<short>("Type")
-                        .HasColumnType("smallint");
+                    b.Property<int>("Type")
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
